rejigger structure of URL's and Views
