+++ readthedocs/core/urls/single_version.py
